lvl 1

+++ src/main/java/student_jekaterina_soldatova/lesson11/level_1_intern/SafeInteractable.java
@@ -5,7 +5,7 @@ class SafeInteractable {
     private double moneyInSafe;
     private String password;
 
-    public SafeInteractable(String password, int moneyInSafe) {
+    public SafeInteractable(String password, double moneyInSafe) {
         this.password = password;
         this.moneyInSafe = moneyInSafe;
     }
